+++ crud/delete.lua
@@ -42,24 +42,31 @@ end
 --
 -- @tparam ?number opts.timeout
 --  Function call timeout
+-- @tparam ?tuples_tomap opts.tuples_tomap
+--  defines type of returned result as map or tuple, default true
 --
--- @return[1] object
+-- @return[1] object / tuple
 -- @treturn[2] nil
 -- @treturn[2] table Error description
 --
 function delete.call(space_name, key, opts)
     checks('string', '?', {
         timeout = '?number',
+        tuples_tomap = '?boolean',
     })
 
     opts = opts or {}
-
+    if opts.tuples_tomap == nil then
+        opts.tuples_tomap = true
+    end
 
     local space = utils.get_space(space_name, vshard.router.routeall())
     if space == nil then
         return nil, DeleteError:new("Space %q doesn't exists", space_name)
     end
+    local space_format = space:format()
 
+    -- compute default buckect_id
     if box.tuple.is(key) then
         key = key:totable()
     end
@@ -80,7 +87,11 @@ function delete.call(space_name, key, opts)
     end
 
     local tuple = results[replicaset.uuid]
-    local object, err = utils.unflatten(tuple, space:format())
+    if opts.tuples_tomap == false then
+        return tuple
+    end
+
+    local object, err = utils.unflatten(tuple, space_format)
     if err ~= nil then
         return nil, DeleteError:new("Received tuple that doesn't match space format: %s", err)
     end