+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * replace
  * upsert
 
+* Support tuples as a result and input type
+
 ## [0.1.0] - 2020-09-23
 
 ### Added